column-name now only caches name when statement is running
Additionally we invalidate the column name cache on (reset).
While statement is not running, answers come directly from the library.
Alternatively, we could always cache but just invalidate when
(step)ping for the first time; even if the schema changes between
prepare and step, it cannot be observed until (step).  The latter
is arguably cleaner--invalidation can be removed from (reset) and
column-names would be simplified--but step would be more complex.