Remove statement from hash-table then set pointer during finalize-transient
If we change the pointer in the statement object, when we then pass it
to hash-table-delete! it won't match and therefore it'll hang around in
the hash-table until the end of the transaction. If the transaction then
fails we'll attempt to reset that statement and that will fail, causing
an exception which will mask the exception that caused the transaction
to abort.

Therefore, we have to remove it from the hash-table first and then set
the pointer to #f.

Signed-off-by: Andy Bennett <andyjpb@register-dynamics.co.uk>