Fix system SQLite3 version detection.
When the compilation would fail, the "use internal sqlite3" script
would stop and have a nonzero exit status.

To prevent any random failure from making the system assume we should
use the system version (which is exactly the wrong default
assumption), the check is inverted: when the script exits with a zero
status, we use the system version, otherwise we use the shipped one.