For system sqlite3, make chicken-sqlite3 a thin wrapper around sqlite3
On macosx the "sqlite3_load_extension" API is removed because macos is silly.
The built-in shell.c assumes the API is available, so compilation against
the system sqlite3 fails.

Generally, compiling a random version of shell.c against another sqlite3
library is not guaranteed to work.

This patch installs "chicken-sqlite3" unconditionally -- if using the
system sqlite3, it's just a wrapper script that calls sqlite3.

chicken-sqlite3 is provided as a debugging convenience for the user,
so if this isn't exactly the right shell or if calling the shell fails,
it's not the end of the world.