Add nested transaction support to active-statements list
This patch makes the necessary changes to the active-statement list
handling code to keep track of nested transactions and ensure that the
correct statements are reset at the correct time.

We currently have a simple list of active statements. When we're in
autocommit mode this list only ever contains a single statement. When
we're inside a transaction, several statements can build up in this
list.

When the transaction fails, we need to reset all the statements in that
transaction. This corresponds to all the statements in that list. If the
transaction succeeds then we allow the statements to remain in whatever
state they find themselves.

In order to support SQLite's SAVEPOINT style of transaction, we need to
be able to keep track of which statements were started in each
transaction so that, if that transaction fails, we know which ones to
reset. Statements started in an outer transaction need to be left alone
and we leave it up to SQLite to ensure that these statements return to
the correct place in their execution.

What happens here is not well specified by the SQLite documentation.

https://sqlite.org/lang_savepoint.html

"The ROLLBACK TO command reverts the state of the database back to what
it was just after the corresponding SAVEPOINT. Note that unlike that
plain ROLLBACK command (without the TO keyword) the ROLLBACK TO command
does not cancel the transaction. Instead of cancelling the transaction,
the ROLLBACK TO command restarts the transaction again at the beginning.
All intervening SAVEPOINTs are canceled, however."

"restarts the transaction again at the beginning" suggests that outer
statements should return to where they were when the inner transaction
was started.

However,

"The ROLLBACK command with a TO clause rolls back transactions going
backwards in time back to the most recent SAVEPOINT with a matching
name. The SAVEPOINT with the matching name remains on the transaction
stack, but all database changes that occurred after that SAVEPOINT was
created are rolled back."

...is silent on the subject of statements that are reading rather than
writing.

This patch assumes that statements in parent transactions do not require
any attention from this egg  when an inner transaction rolls back.

When transactions succeed, we fold their list of statements into the
parent transaction's list of statements.

The existing `commit` and `rollback` procedures work with the existing
transaction model. When they are called, all nested transactions are
committed or rolled-back.

Signed-off-by: Andy Bennett <andyjpb@register-dynamics.co.uk>