SAVEPOINT Transactions
This patch provides preliminary support for SQLite's nested transaction model.

This patch provides the `with-savepoint-transaction` API that is similar
to the exisiting `with-transaction` APIs. The supplied thunk must always
exit regularly exactly once and thunks from inner calls to
`with-savepoint-transaction` must exit before thunk from outer calls. A
return value of #f signals that the egg should rollback the transaction
to the SAVEPOINT. Any other return value signals that the egg should
release (i.e commit) the transaction.

We try to roll it into the existing APIs that the egg provides for
transactions however, some caveats are present.

This egg tries to provide a high level API for using the SQLite database
engine. However, it is possible for the user to use their own SQL
statements that "begin;", "commit;" or "rollback;" a transaction. This
makes it difficult for the egg to keep track of the transaction state.
With regular, unnestable, transactions it is difficult to cause problems
and the consequences are usually minor. With SAVEPOINT transactions, if
the API is not strictly followed, it is possible for statements to be
incorrectly reset which may affect program correctness.

The SQLite3 C API only provides callback hooks for committing or
rolling-back; not for begining transactions. Therefore, we do as best as
we can at keeping track of the state of the transaction stack. In
particular, the following caveats apply:

 + If the user executes "begin;", then use of the `commit` or `rollback`
   procedure will try to issue a warning. (This is implemented in the
   active-statement list patch.)

 + If the user executes "rollback;" then the egg will not properly reset
   the in-progress statements. (This is the egg's existing behaviour.)

 + If the user does not exit SAVEPOINT transactions in the order they
   were entered (i.e by executing their own "rollback;", "commit;" or
   "rollback to ...;" statement or by modifying the default control flow
   of the `with-savepoint-transaction` API) then the state of the egg
   will diverge from the state of the SQLite database engine.

 + If the user calls the `commit` or `rollback` procedures whilst inside
   a SAVEPOINT transaction then the state of the egg will diverge from
   the state of the SQLite database engine. However, the egg will try to
   signal an error if the regular control flow of the
   `with-savepoint-transaction` API concludes.

At the points where the egg tries to signal an error or issue a warning,
it may be unable to detect the situation and therefore these errors and
warnings are on a best effort basis. Incorrect use of the API in the
past may cause these errors and warnings to appear later in unrelated
and correct parts of the code.

Signed-off-by: Andy Bennett <andyjpb@register-dynamics.co.uk>